feat(android): AuthViewModel i cuvanje tokena preko DataStore-a
TokenStore perzistira token + korisnika; AuthViewModel izlaze StateFlow stanje (korisnik/token/greska/ucitava) i login/register/osvezi/logout preko
Retrofit-a.
Model: brojBodova i popust prebaceni na Double (backend salje decimalno).

+++ glumacplusandroid/app/src/main/java/com/example/glumac_plus_android/data/model/Korisnik.kt
@@ -10,8 +10,8 @@ data class Korisnik(
     val id: Long,
     val ime: String,
     val email: String,
-    val uloga: String? = null,        // ? + = null → polje sme da fali (nullable, kao "?" u TS)
-    val brojBodova: Int = 0,
+    val uloga: String? = null,
+    val brojBodova: Double = 0.0,
     val loyaltyNivo: String? = null
 )
 
@@ -19,7 +19,7 @@ data class Korisnik(
 data class LoyaltyProgram(
     val id: Long,
     val nivo: String,
-    val popust: Int,
+    val popust: Double,
     val pragBodova: Int
 )
